Fixing bug that would cause check-reserve-update-frequency tests to display warning 'Jest did not exit one second after the test run has completed'.
